# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Micro Wallet Documentation",
        "description": "Software for manage wallets and transictions by users",
        "contact": {
            "email": "menezes.adolfo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "API Documentation for Micro Wallet"
        }
    ],
    "paths": {
        "/transactions/byWallet/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get list of transactions by wallet id",
                "description": "Returns transactions by wallet id",
                "operationId": "showTransactionsByWalletId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "wallet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "transactions by wallet id",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "deposits": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "wallet_payer_id": {
                                                                    "type": "number"
                                                                },
                                                                "amount": {
                                                                    "type": "number"
                                                                },
                                                                "description": {
                                                                    "type": "string"
                                                                },
                                                                "type": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                },
                                                                "notify": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "withdraws": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "wallet_payer_id": {
                                                                    "type": "number"
                                                                },
                                                                "amount": {
                                                                    "type": "number"
                                                                },
                                                                "description": {
                                                                    "type": "string"
                                                                },
                                                                "type": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                },
                                                                "notify": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "transfers": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "number"
                                                                },
                                                                "wallet_payer_id": {
                                                                    "type": "number"
                                                                },
                                                                "wallet_payee_id": {
                                                                    "type": "number"
                                                                },
                                                                "amount": {
                                                                    "type": "number"
                                                                },
                                                                "description": {
                                                                    "type": "string"
                                                                },
                                                                "type": {
                                                                    "type": "string"
                                                                },
                                                                "status": {
                                                                    "type": "string"
                                                                },
                                                                "notify": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "List transaction by id",
                "description": "Return data of transaction by id",
                "operationId": "getTransactionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of transaction by id",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "wallet_payer_id": {
                                                    "type": "number"
                                                },
                                                "wallet_payee_id": {
                                                    "type": "number"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "notify": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/transactions": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create transactions",
                "description": "Return create new transaction",
                "operationId": "storeTransactions",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "wallet_payer_id": {
                                        "type": "number"
                                    },
                                    "wallet_payee_id": {
                                        "type": "number"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of new transaction",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "wallet_payer_id": {
                                                    "type": "number"
                                                },
                                                "wallet_payee_id": {
                                                    "type": "number"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "notify": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}